pyproject.toml
[project]
name = "multi-resolution-crowd-counting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "einops>=0.8.2",
    "ftfy>=6.3.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "pydantic-settings>=2.0",
    "regex>=2026.4.4",
    "scipy>=1.17.1",
    "tensorboardx>=2.6.5",
    "timm>=1.0.26",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
]
